Implement StatModifier and fix some problems with Stats.
Stats now has a builder class so that it is easier to cleanly create
a new Stats object. Additionally, removed some of the unnecessary
properties and made them just methods. There was no reason to have
members to store the derived stats when the formulas for them are
so simple.